requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
h5py>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Logging & Progress
tqdm>=4.65.0
tensorboard>=2.12.0

# Utilities
scikit-learn>=1.2.0
